requirements.txt
gradio>=3.40
python-docx>=0.8.11
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=0.21.0
